Clean up DeepPtr

+++ ASL/ASLParser.cs
@@ -27,8 +27,8 @@ namespace LiveSplit.ASL
                 var type = (String)childNodes[0].Token.Value;
                 var identifier = (String)childNodes[1].Token.Value;
                 var module = (String)childNodes[3].Token.Value;
-                var moduleBase = childNodes[5].ChildNodes.Select(x => (long)(int)x.Token.Value).First();
-                var offsets = childNodes[5].ChildNodes.Skip(1).Select(x => (long)(int)x.Token.Value).ToArray();
+                var moduleBase = childNodes[5].ChildNodes.Select(x => (int)x.Token.Value).First();
+                var offsets = childNodes[5].ChildNodes.Skip(1).Select(x => (int)x.Token.Value).ToArray();
 
                 // this assumes offsets will never be bigger than 32 bit!
